ingesta/requirements.txt
fastapi
uvicorn
aiohttp
lxml
beautifulsoup4
apscheduler
aiomysql
python-dotenv
cryptography
